+++ app/models/facilities_management/rm6378/journey/annual_contract_value.rb
@@ -3,12 +3,12 @@ module FacilitiesManagement
     class Journey::AnnualContractValue
       include Steppable
 
-      attribute :annual_contract_value, Numeric
+      attribute :annual_contract_value, :numeric
 
       attribute :contract_start_date_dd
       attribute :contract_start_date_mm
       attribute :contract_start_date_yyyy
-      attribute :estimated_contract_duration, Numeric
+      attribute :estimated_contract_duration, :numeric
       attribute :private_finance_initiative
 
       validates :annual_contract_value, numericality: { only_integer: true, greater_than: 0, less_than: 1_000_000_000_000 }